scripts/requirements.txt
requests
python-dotenv
dnspython